Add files via upload
Standard Input in Dart (stdin):
The standard input stream reads data both synchronously and asynchronously from the keyboard. 

In Dart programming language, .readLineSync( ) function is used to accept input from the user via console. The definition of this function is stored in the ‘dart:io’ library; therefore, to use its functions and objects, you need to import this library in your code. If you fail to do so, the compiler will report an error, and the program will not execute.